+++ backend/canisters/community/api/src/queries/selected_initial.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{AccessRules, CommunityMember, EventIndex, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
+use types::{CommunityMember, EventIndex, Rules, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -20,8 +20,8 @@ pub struct SuccessResult {
     pub members: Vec<CommunityMember>,
     pub blocked_users: Vec<UserId>,
     pub invited_users: Vec<UserId>,
-    // TODO: remove this field once the website is using `access_rules` instead
-    pub rules: AccessRules,
-    pub access_rules: VersionedRules,
+    // TODO: remove this field once the website is using `chat_rules` instead
+    pub rules: Rules,
+    pub chat_rules: VersionedRules,
     pub user_groups: Vec<UserGroupDetails>,
 }