+++ backend/canisters/group/api/src/updates/convert_into_community.rs
@@ -1,10 +1,10 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{AccessRules, ChannelId, CommunityId, CommunityPermissions};
+use types::{ChannelId, CommunityId, CommunityPermissions, Rules};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub rules: AccessRules,
+    pub rules: Rules,
     pub permissions: Option<CommunityPermissions>,
     pub primary_language: Option<String>,
     pub history_visible_to_new_joiners: bool,