+++ backend/canisters/group/impl/src/queries/selected_initial.rs
@@ -27,7 +27,7 @@ fn selected_initial_impl(state: &RuntimeState) -> Response {
                 .copied()
                 .collect(),
             rules: chat.rules.clone().into(),
-            access_rules: chat.rules.clone().into(),
+            chat_rules: chat.rules.clone().into(),
         })
     } else {
         CallerNotInGroup